requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
fastapi==0.134.0
filelock==3.24.3
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.65.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.10
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.7.4
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.47
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
